Feature/convolutional 2d model support (#22)
* Removed buggy 440Hz file and added 1kHz test sound file.
* Every sound sample is read as float32
* Now by default periodogram and spectrogram features are calculated based on float values (disable raw loggic)

+++ features/sound_dataset.py
@@ -40,7 +40,7 @@ def _read_samples(filename: Path, raw: bool = False) -> (np.ndarray, int):
         sound_samples = sound_samples.sum(axis=1) // 2
 
     if not raw and sound_samples.dtype.kind != 'f':
-        sound_samples = _pcm2float(sound_samples, dynamic_type='float64')
+        sound_samples = _pcm2float(sound_samples, dynamic_type='float32')
 
     return sound_samples, sampling_rate
